+++ core/src/main/scala/org/typelevel/vault/Locker.scala
@@ -22,29 +22,35 @@
 package org.typelevel.vault
 
 import cats.implicits._
-import org.typelevel.unique.Unique
+import cats.effect.kernel.Unique
 
 /**
- * Locker - A persistent store for a single value.
- * This utilizes the fact that a unique is linked to a type.
- * Since the key is linked to a type, then we can cast the
- * value to Any, and join it to the Unique. Then if we
- * are then asked to unlock this locker with the same unique, we
- * know that the type MUST be the type of the Key, so we can
- * bring it back as that type safely.
+ * Locker - A persistent store for a single value. This utilizes the fact that a unique is linked to a type. Since the
+ * key is linked to a type, then we can cast the value to Any, and join it to the Unique. Then if we are then asked to
+ * unlock this locker with the same unique, we know that the type MUST be the type of the Key, so we can bring it back
+ * as that type safely.
  */
-final class Locker private (private val unique: Unique, private val a: Any) {
+final class Locker private (private val unique: Unique.Token, private val a: Any) {
 
   /**
-   * Retrieve the value from the Locker. If the reference equality
-   * instance backed by a `Unique` value is the same then allows
-   * conversion to that type, otherwise as it does not match
-   * then this will be `None`
+   * Retrieve the value from the Locker. If the reference equality instance backed by a `Unique` value is the same then
+   * allows conversion to that type, otherwise as it does not match then this will be `None`
    *
-   * @param k The key to check, if the internal Unique value matches
-   * then this Locker can be unlocked as the specifed value
+   * @param k
+   *   The key to check, if the internal Unique value matches then this Locker can be unlocked as the specifed value
    */
-  def unlock[A](k: Key[A]): Option[A] = Locker.unlock(k, this)
+  def unlock[A](k: LookupKey[A]): Option[A] =
+    if (k.unique === unique) Some(k.out(a.asInstanceOf[k.I]))
+    else None
+
+  /**
+   * Retrieve the value from the Locker. If the reference equality instance backed by a `Unique` value is the same then
+   * allows conversion to that type, otherwise as it does not match then this will be `None`
+   *
+   * @param k
+   *   The key to check, if the internal Unique value matches then this Locker can be unlocked as the specifed value
+   */
+  private[vault] def unlock[A](k: Key[A]): Option[A] = unlock(k: LookupKey[A])
 }
 
 object Locker {
@@ -52,20 +58,23 @@ object Locker {
   /**
    * Put a single value into a Locker
    */
-  def lock[A](k: Key[A], a: A): Locker = new Locker(k.unique, a.asInstanceOf[Any])
+  def apply[A](k: InsertKey[A], a: A): Locker = new Locker(k.unique, k.in(a))
 
   /**
-   * Retrieve the value from the Locker. If the reference equality
-   * instance backed by a `Unique` value is the same then allows
-   * conversion to that type, otherwise as it does not match
-   * then this will be `None`
-   *
-   * @param k The key to check, if the internal Unique value matches
-   * then this Locker can be unlocked as the specifed value
-   * @param l The locked to check against
+   * Put a single value into a Locker
    */
-  def unlock[A](k: Key[A], l: Locker): Option[A] =
-    // Equality By Reference Equality
-    if (k.unique === l.unique) Some(l.a.asInstanceOf[A])
-    else None
+  @deprecated("Use Locker(k, a)", since = "3.1.0")
+  def lock[A](k: Key[A], a: A): Locker = Locker(k, a)
+
+  /**
+   * Retrieve the value from the Locker. If the reference equality instance backed by a `Unique` value is the same then
+   * allows conversion to that type, otherwise as it does not match then this will be `None`
+   *
+   * @param k
+   *   The key to check, if the internal Unique value matches then this Locker can be unlocked as the specifed value
+   * @param l
+   *   The locked to check against
+   */
+  @deprecated("Use l.unlock(k)", since = "3.1.0")
+  def unlock[A](k: Key[A], l: Locker): Option[A] = l.unlock(k)
 }