+++ CHANGELOG.md
@@ -6,7 +6,17 @@ This file summarizes **notable** changes for each release, but does not describe
 
 ## <a name="Unreleased"></a>Unreleased Changes
 
-## <a name="2.1.1"></a>New and Noteworthy for Version 2.1.0
+## <a name="3.0.0-RC1"></a>New and Noteworthy for Version 3.0.0-RC1
+
+* Built for cats-effect-3.0.0-RC1
+* Uses Cats-Effect-3's `Unique` instead of our internal one
+
+## <a name="3.0.0-M1"></a>New and Noteworthy for Version 3.0.0-M1
+
+* Built for cats-effect-3.0.0-M5
+* Internal, private copy of Unique
+
+## <a name="2.1.7"></a>New and Noteworthy for Version 2.1.7
 
 * Add support for Dotty 3.0.0-RC1
 * Drop support for Dotty 3.0.0-M2